gpu: nvgpu: acr: cast assignments of sizeof to u32
This change fixes a number of This is a MISRA 10.3 rule violation due to
the implicit casts of sizeof() to u32's. This change adds u32 casts to
each of these violations. This should be safe because a 4GB type size
would be very unlikely in this driver.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/1959428
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/pmu/acr_gm20b.c
@@ -565,7 +565,7 @@ int gm20b_pmu_populate_loader_cfg(struct gk20a *g,
 	ldr_cfg->argc = 1;
 	ldr_cfg->argv = addr_args;
 
-	*p_bl_gen_desc_size = sizeof(struct loader_config);
+	*p_bl_gen_desc_size = (u32)sizeof(struct loader_config);
 	g->acr.pmu_args = addr_args;
 	return 0;
 }
@@ -625,7 +625,7 @@ int gm20b_flcn_populate_bl_dmem_desc(struct gk20a *g,
 	ldr_cfg->data_dma_base = addr_data;
 	ldr_cfg->data_size = desc->app_resident_data_size;
 	ldr_cfg->code_entry_point = desc->app_imem_entry;
-	*p_bl_gen_desc_size = sizeof(struct flcn_bl_dmem_desc);
+	*p_bl_gen_desc_size = (u32)sizeof(struct flcn_bl_dmem_desc);
 	return 0;
 }
 
@@ -676,7 +676,8 @@ static void lsfm_init_wpr_contents(struct gk20a *g, struct ls_flcn_mgr *plsfm,
 	while (pnode != NULL) {
 		/* Flush WPR header to memory*/
 		nvgpu_mem_wr_n(g, ucode, i * sizeof(pnode->wpr_header),
-				&pnode->wpr_header, sizeof(pnode->wpr_header));
+				&pnode->wpr_header,
+				(u32)sizeof(pnode->wpr_header));
 
 		nvgpu_pmu_dbg(g, "wpr header");
 		nvgpu_pmu_dbg(g, "falconid :%d",
@@ -692,7 +693,8 @@ static void lsfm_init_wpr_contents(struct gk20a *g, struct ls_flcn_mgr *plsfm,
 
 		/*Flush LSB header to memory*/
 		nvgpu_mem_wr_n(g, ucode, pnode->wpr_header.lsb_offset,
-				&pnode->lsb_header, sizeof(pnode->lsb_header));
+				&pnode->lsb_header,
+				(u32)sizeof(pnode->lsb_header));
 
 		nvgpu_pmu_dbg(g, "lsb header");
 		nvgpu_pmu_dbg(g, "ucode_off :%x",
@@ -743,7 +745,7 @@ static void lsfm_init_wpr_contents(struct gk20a *g, struct ls_flcn_mgr *plsfm,
 	nvgpu_mem_wr_n(g, ucode,
 			plsfm->managed_flcn_cnt * sizeof(struct lsf_wpr_header),
 			&last_wpr_hdr,
-			sizeof(struct lsf_wpr_header));
+			(u32)sizeof(struct lsf_wpr_header));
 }
 
 /*!
@@ -961,7 +963,7 @@ static int lsf_gen_wpr_requirements(struct gk20a *g, struct ls_flcn_mgr *plsfm)
 		wpr_offset = ALIGN(wpr_offset,
 			LSF_LSB_HEADER_ALIGNMENT);
 		pnode->wpr_header.lsb_offset = wpr_offset;
-		wpr_offset += sizeof(struct lsf_lsb_header);
+		wpr_offset += (u32)sizeof(struct lsf_lsb_header);
 
 		/* Align, save off, and include the original (static)
 		ucode image size */
@@ -1385,7 +1387,7 @@ static void gm20b_acr_default_sw_init(struct gk20a *g, struct hs_acr *hs_acr)
 
 	/* bootlader interface used by ACR HS bootloader*/
 	hs_acr->ptr_bl_dmem_desc = &hs_acr->bl_dmem_desc;
-	hs_acr->bl_dmem_desc_size = sizeof(struct flcn_bl_dmem_desc);
+	hs_acr->bl_dmem_desc_size = (u32)sizeof(struct flcn_bl_dmem_desc);
 
 	/* set on which falcon ACR need to execute*/
 	hs_acr->acr_flcn = &g->pmu_flcn;
