+++ drivers/gpu/nvgpu/common/pmu/acr_gv11b.c
@@ -152,7 +152,7 @@ static void gv11b_acr_default_sw_init(struct gk20a *g, struct hs_acr *hs_acr)
 	hs_acr->acr_fw_name = HSBIN_ACR_UCODE_IMAGE;
 
 	hs_acr->ptr_bl_dmem_desc = &hs_acr->bl_dmem_desc_v1;
-	hs_acr->bl_dmem_desc_size = sizeof(struct flcn_bl_dmem_desc_v1);
+	hs_acr->bl_dmem_desc_size = (u32)sizeof(struct flcn_bl_dmem_desc_v1);
 
 	hs_acr->acr_flcn = &g->pmu_flcn;
 	hs_acr->acr_flcn_setup_hw_and_bl_bootstrap =
